[tests] Add 'CheckedThread' as a common test utility
